+++ src/main/kotlin/domain/Service.kt
@@ -2,4 +2,4 @@ package domain
 
 import java.math.BigDecimal
 
-data class Service (val id: Int, val employee: Employee, val skills: Skills, val price: BigDecimal)
+data class Service (val id: Int, val employee: Employee, val skills: Skills, val price: Int)
